requirements.txt
google-adk
google-genai
openai
PyYAML
python-dotenv
pytest
kivy